+++ system-tests/project-fixtures/runner-specs/cypress/support/component.js
@@ -1,4 +1,3 @@
-import { mount, unmount } from 'cypress/react18'
+import { mount } from 'cypress/react18'
 
 Cypress.Commands.add('mount', mount)
-Cypress.Commands.add('unmount', unmount)